As requested, moved InOutOctree up one level; fixed old typo in Sidre figure
